Removed Cholesky(normal=True, ...) (#12)
I realised that the implementation was subpar, and the result just not that useful.

First of all, the implementation was subpar because it directly materialised A.T @ A and then computed cho_factor(A.T @ A). However it is more efficient+stable to just just compute the QR decomposition of A, for which A.T @ A = R.T @ Q.T @ Q @ R = R.T @ R so that L = R.T is a lower-triangular Cholesky decmposition of A.T @ A.

And at that point, why not just use a QR solver directly? (There are some nitty differences around the edges: Cholesky only needs R so it can skip the Householder product for computing Q, but in return it requires two downstream triangular solves instead of just one.) The extra option isn't really worth the maintenance burden. Thus, removed.