Removed non-iterative solvers.
These were originally introduce to handle LearningRate, which was a
minimiser that worked by taking a single step.

However we've now separated out searches+descents into their own thing,
not dependent on the rest of the solver infrastructure. So we have no
more need for non-iterative solvers.

We could still reintroduce them at some later date, if we find that we
have some exotic non-iterative solver we'd like to implement.